requirements.txt
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.2
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
kombu==5.5.3
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
prompt_toolkit==3.0.51
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
redis==6.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
